requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0

# Real-time Communication
Flask-SocketIO==5.3.6
python-socketio==5.11.0
eventlet==0.35.2

# API & Validation
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
marshmallow==3.20.2
email-validator==2.1.0

# Payment Integration
stripe==7.11.0

# Digital Signature & Document Processing
docusign-esign==3.26.0
PyPDF2==3.0.1
cryptography==41.0.7

# AI/ML Libraries
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
tensorflow==2.18.0

# Google APIs
googlemaps==4.10.0
google-generativeai==0.3.2

# Image Processing
Pillow==10.2.0
python-magic==0.4.27

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1

# Development & Testing
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0

# Monitoring & Logging
python-json-logger==2.0.7

# Rate Limiting
Flask-Limiter==3.5.1

